agents/voice-synthesis/requirements.txt
# Pin specific versions to avoid dependency resolution issues
google-cloud-aiplatform==1.125.0
google-api-python-client==2.184.0
fastapi==0.115.0
click==8.1.8
authlib==1.5.1
pydantic==2.9.2
pydantic-core==2.23.4
uvicorn==0.34.0
deprecated==1.2.14

# Then install google-adk
google-adk==0.1.0

# Project-specific dependencies
google-cloud-firestore==2.14.0
google-cloud-storage==2.18.0
google-cloud-texttospeech==2.16.0
python-dotenv==1.0.0
